requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
psycopg==3.2.3
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2
